rag_framework/requirements.txt
# Core environment management
python-dotenv>=1.0.0

# Vector storage
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0
transformers==4.57.1

# LangChain orchestration
langchain>=1.0.6
langchain-community>=0.2.0
langchain-text-splitters>=1.0.0
langchain-huggingface==1.0.1

# Groq LLM client
groq>=0.5.0

# Experiment tracking
mlflow>=2.14.0

# Evaluation
ragas>=0.1.0
pandas>=2.0.0

# Streamlit (if you want a UI)
streamlit>=1.39.0

# PDF and document parsing (optional, for ingestion)
PyPDF2>=3.0.0
PyMuPDF
unstructured>=0.15.0
